docker/alarm_collector/requirements.txt
paho-mqtt
pymongo
uvicorn
python-dotenv
line-bot-sdk